Feat: support resource topology for endpoints
