+++ pkg/stdlib/pkgs/query.cue
@@ -50,6 +50,9 @@
 		deployVersion?:   string
 		revision?:        string
 		latest?:          bool
+		resourceTree?: {
+			...
+		}
 	}]
 	...
 }
